Replace boost::regex with RE2 regex in url_extract_parameter and fb_url_extract_parameter Presto function in Velox (#14681)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/14681

Some implementations of functions in Velox rely on boost::regex as regular expression engine. That is not recommended. RE2 is the preferred engine.

boost relies on backtracking and it can lead to significantly slower performance compared to RE2. RE2 has O(n) complexity for a given string of length n, as it uses finite-state machines

Differential Revision: D81494279